[Code] Change `%assert-list-length` into macro
To allow codes using this feature not load "destructuring-bind" once it
is compiled.

As being a function and it must be called at run-time, it must appear
in expansion. Leaving a symbol from "destructuring-bind" package is
something I don't appreciate, 'coz it requires destructuring-bind to be
loaded even if it is compiled(expanded) already.